valid test configuration - avoid library build before test
