+++ package.json
@@ -9,21 +9,25 @@
     "type": "git",
     "url": "git@github.com:backtrace-labs/backtrace-node.git"
   },
+  "scripts": {
+    "test": "mocha --require ts-node/register -r ./tsconfig.json --project tsconfig.json test/**/*.ts",
+    "build": "tsc"
+  },
   "dependencies": {
     "axios": "^0.18.0",
     "form-data": "^2.3.3",
     "node-machine-id": "^1.1.10",
-    "source-scan": "~1.0.1",
-    "streamsink": "~1.2.0"
-  },
-  "scripts": {
-    "test": "mocha",
-    "build": "tsc"
+    "source-scan": "~1.0.1"
   },
   "devDependencies": {
+    "@types/chai": "^4.1.7",
+    "@types/expect": "^1.20.4",
     "@types/form-data": "^2.2.1",
+    "@types/mocha": "^5.2.6",
     "@types/node": "^11.11.4",
     "chai": "~4.2.0",
-    "mocha": "~5.2.0"
+    "mocha": "^5.2.0",
+    "ts-node": "^8.0.3",
+    "typescript": "^3.3.4000"
   }
 }
