--- source/model/backtraceData.ts
@@ -1,16 +1,0 @@
-import { ISourceCode } from './sourceCode';
-
-export interface BacktraceData {
-  uuid: string;
-  timestamp: number;
-  lang: string;
-  langVersion: string;
-  agent: string;
-  agentVersion: string;
-  mainThread: string;
-  attributes: object;
-  annotations: object;
-  threads: object;
-  classifiers: string[];
-  sourceCode: { [index: string]: ISourceCode };
-}
